Preparation for block lighting.
Changed block id array from 4 bytes per block to 2 bytes per block. Added light value array, 4 bytes per block. Total now 6 bytes per block. Memory usage up from about 2.75 GB to about 4 GB. Lighting plan is 8 half-byte channels: Sun, Red, Green, Blue, Yellow, Magenta, Cyan, White.

+++ src/minusk/tiletech/utils/Util.java
@@ -12,7 +12,7 @@ public class Util {
 		return xyz < 0 ? (xyz-31) / 32 : xyz / 32;
 	}
 	
-	public static boolean contains(int v, int[] a) {
+	public static boolean contains(short v, short[] a) {
 		for (int c : a)
 			if (v == c)
 				return true;
