+++ src/minusk/tiletech/world/tiles/standard/StandardTransparentTile.java
@@ -4,7 +4,7 @@ package minusk.tiletech.world.tiles.standard;
  * Created by MinusKelvin on 2/4/16.
  */
 public class StandardTransparentTile extends StandardSolidTile {
-	public StandardTransparentTile(int id, int top, int bottom, int east, int west, int north, int south) {
+	public StandardTransparentTile(short id, int top, int bottom, int east, int west, int north, int south) {
 		super(id, top, bottom, east, west, north, south);
 	}
